Build system: Change hard coded library path with CMake variable.
Removes the hard coded library path `lib`. Replaces it with
`CMAKE_INSTALL_LIBDIR`. This also sets the same path in the pkg-config
file.

The path defaults to 'lib', but it can be overridden if specifying
`-DCMAKE_INSTALL_LIBDIR=your_relative_libdir` during CMake configuration.
The variable is also available for manual change when configuring with
`ccmake`.

The `CMAKE_INSTALL_LIBDIR` is included with GNUInstallDirs.