blosc/blosc2.c: optimize the mutex lock with spin lock
The mutex lock cost lots of cycles in t_blosc_do_job when the thread
number is large. The updating to global values of protected part is
not that heavy. It will save lots of cycles if we replace the mutex
lock with spin lock.

Apply the patch and test the c-blosc2/bench.
$ ./bench/b2bench blosclz noshuffle single 160 83886080 4
Score gain: 7.4%

$ ./bench/b2bench blosclz noshuffle single 64 83886080 4
Score gain: 3.5%

$ ./bench/b2bench blosclz noshuffle single 32 83886080 4
Score gain: 1.9%

CPU: ICX 8380 x 2 sockets
Core number: 160/64/32 threads