b2nd: Use const where possible.
The current API doesn't use `const` on pointers when treating them as
read-only, which makes the API difficult to use in a const-safe way. It
also makes it difficult for users to understand the semantics.

Note that casts had to be used internally in a few places because of the
use of functions which handle both read and write paths depending on a
flag.