Stop supporting ancient versions of Visual Studio
This removes support for _MSC_VER <= 1600, that is Visual Studio 2010
(version 10.0) or earlier.

We expect ends users will run Visual Studio 2012 (version 11.0) or later,
which means _MSC_VER >= 1700. This is not to say Studio 2012 (version 11.0)
will work our of the box, for lack of testing, just that anything
earlier will not work.