Fix: CMake Config Version
When writing to CMake, we cannot append the suffix string
in the version file, i.e., have to omit `.dev`.

The reason for that is that `VERSION_GREATER_EQUAL` is not that
flexible: https://cmake.org/cmake/help/latest/variable/CMAKE_VERSION.html
and downstream users comparing against the version will get a
`FALSE` result comparing things like
```cmake
if (Blosc2_VERSION VERSION_GREATER_EQUAL 2.10.1)
```
if `Blosc2_VERSION` is set to `2.10.2.dev`.