fastcopy: use memcpy instead of open-coding it
This is the same thing and will help even less-intelligent compilers. No point
in open-coding it.

Keep the various copy_*_bytes helpers for compatibility. There's no real cost
to this because they're all inline.

On the other hand, we probably saved a bit of compile-time here because we're
dropping a bunch of inlines because of the indirection (consider copy_7_bytes).

Signed-off-by: Sam James <sam@gentoo.org>