Use the NEON bitshuffle code in the bitshuffle project.
Although it is slower that our existing code, and of course,
slower than the generic code (at least on an ARM M1), at least
this is completely correct, even for typesizes 1, 2 and 4 (the
previous code was not).

Also, syncing sources with the bitshuffle project makes things
easier when porting docs and fixes from there.