Fix public include directories for blosc2 targets:
The target_include_directories interface is given in blosc2 subdirectory as ${CMAKE_CURRENT_SOURCE_DIR},
which makes this directory automatically included when linking the blosc2 target in CMake. However,
the blosc2 include directory is in the parent ../include directory, so the correct way to include it could be
${CMAKE_CURRENT_SOURCE_DIR}/../include . Moreover, ${CMAKE_CURRENT_SOURCE_DIR} exposes internal headers
instead of the public headers, so we may want to turn the original include directory into a private include.